+++ src/IconCheckbox.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+import IconLayout from './IconLayout';
+import Checkbox, { CHECKBOX_BUTTON } from './Checkbox';
+
+function IconCheckbox(props) {
+    const buttonWithStatus = <IconLayout icon={CHECKBOX_BUTTON} />;
+
+    return (
+        <Checkbox
+            minified
+            overrideButton={buttonWithStatus}
+            {...props}>
+            <span /> {/* to trick <RowComp> from rendering default content */}
+        </Checkbox>
+    );
+}
+
+export default IconCheckbox;